refactor: split admin keys and dialogs
