+++ src/ui/sidebar/AdminSettingsPanel.tsx
@@ -8,7 +8,6 @@ import {
   createApiKey,
   deleteUser,
   revokeAllUserSessions,
-  deleteApiKey,
   createRole,
   registerUser,
   makeUserAdmin,
@@ -34,29 +33,8 @@ import {
   updateOidcAutoProvision,
   isElectron,
   getUserRoles,
-  assignRoleToUser,
-  removeRoleFromUser,
 } from "@/main-axios";
 import type { ApiKey, CreatedApiKey, UserRole } from "@/main-axios";
-import { Button } from "@/components/button";
-import { Input } from "@/components/input";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogDescription,
-} from "@/components/dialog";
-import {
-  AlertCircle,
-  Copy,
-  Eye,
-  EyeOff,
-  Network,
-  Plus,
-  RefreshCw,
-  Trash2,
-} from "lucide-react";
 import type { AdminSection } from "@/types/ui-types";
 import type { Role } from "@/main-axios";
 import {
@@ -68,7 +46,6 @@ import {
 } from "./AdminManagementSections";
 import { toast } from "sonner";
 import { getBasePath } from "@/lib/base-path";
-import { AdminToggle } from "./AdminSettingsShared";
 import {
   AdminDatabaseSection,
   AdminGeneralSettingsSection,
@@ -787,542 +764,59 @@ export function AdminSettingsPanel() {
         handleImportDatabase={handleImportDatabase}
       />
 
-      {/* API Keys */}
-      <AccordionSection
-        label={t("admin.sectionApiKeys")}
-        icon={<Network className="size-3.5" />}
+      <AdminApiKeysSection
         open={openSection === "api-keys"}
         onToggle={() => toggle("api-keys")}
-      >
-        <div className="flex flex-col pt-2">
-          <div className="flex items-center justify-between py-2 border-b border-border">
-            <span className="text-[10px] text-muted-foreground">
-              {t("admin.apiKeysCount", { count: apiKeys.length })}
-            </span>
-            <div className="flex items-center gap-1">
-              <Button
-                variant="ghost"
-                size="icon"
-                className="size-6 text-muted-foreground hover:text-foreground"
-                onClick={loadApiKeys}
-              >
-                <RefreshCw className="size-3" />
-              </Button>
-              <Button
-                variant="outline"
-                size="sm"
-                className="h-6 text-[10px] border-accent-brand/40 text-accent-brand hover:bg-accent-brand/10 hover:text-accent-brand"
-                onClick={() => {
-                  setShowCreateKey((o) => !o);
-                  setCreatedKeyToken(null);
-                }}
-              >
-                <Plus className="size-3" />
-                {t("admin.createRole")}
-              </Button>
-            </div>
-          </div>
-          {showCreateKey && (
-            <div className="flex flex-col gap-2.5 py-3 border-b border-border">
-              <span className="text-[10px] font-semibold uppercase tracking-widest text-muted-foreground">
-                {t("admin.newApiKey")}
-              </span>
-              {createdKeyToken ? (
-                <div className="flex flex-col gap-2">
-                  <span className="text-[10px] text-accent-brand font-semibold">
-                    {t("admin.apiKeyCreatedWarning")}
-                  </span>
-                  <div className="flex items-center gap-2 bg-muted/30 border border-border px-2 py-1.5">
-                    <span className="text-[10px] font-mono flex-1 truncate">
-                      {createdKeyToken}
-                    </span>
-                    <button
-                      onClick={() => {
-                        navigator.clipboard.writeText(createdKeyToken);
-                        toast.info(t("admin.copiedToClipboard"));
-                      }}
-                      className="text-muted-foreground hover:text-accent-brand shrink-0"
-                    >
-                      <Copy className="size-3.5" />
-                    </button>
-                  </div>
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    className="text-xs self-end"
-                    onClick={() => {
-                      setShowCreateKey(false);
-                      setCreatedKeyToken(null);
-                    }}
-                  >
-                    {t("admin.done")}
-                  </Button>
-                </div>
-              ) : (
-                <>
-                  <div className="flex flex-col gap-1">
-                    <label className="text-[10px] font-semibold text-muted-foreground uppercase tracking-widest">
-                      {t("admin.apiKeyName")}{" "}
-                      <span className="text-accent-brand">*</span>
-                    </label>
-                    <Input
-                      placeholder="e.g., CI Pipeline"
-                      value={newKeyName}
-                      onChange={(e) => setNewKeyName(e.target.value)}
-                      className="text-xs"
-                    />
-                  </div>
-                  <div className="flex flex-col gap-1">
-                    <label className="text-[10px] font-semibold text-muted-foreground uppercase tracking-widest">
-                      {t("admin.apiKeyUser")}{" "}
-                      <span className="text-accent-brand">*</span>
-                    </label>
-                    <select
-                      className="px-2 py-1.5 text-xs bg-background border border-border text-foreground outline-none"
-                      value={newKeyUserId}
-                      onChange={(e) => setNewKeyUserId(e.target.value)}
-                    >
-                      <option value="">{t("admin.apiKeySelectUser")}</option>
-                      {users.map((u) => (
-                        <option key={u.id} value={u.id}>
-                          {u.username}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="flex flex-col gap-1">
-                    <label className="text-[10px] font-semibold text-muted-foreground uppercase tracking-widest">
-                      {t("admin.apiKeyExpiresAt")}
-                    </label>
-                    <Input
-                      type="date"
-                      value={newKeyExpiry}
-                      onChange={(e) => setNewKeyExpiry(e.target.value)}
-                      className="text-xs"
-                    />
-                  </div>
-                  <div className="flex justify-end gap-2">
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="text-xs"
-                      onClick={() => setShowCreateKey(false)}
-                    >
-                      {t("common.cancel")}
-                    </Button>
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      className="text-xs border-accent-brand/40 text-accent-brand hover:bg-accent-brand/10 hover:text-accent-brand"
-                      onClick={handleCreateApiKey}
-                      disabled={newKeyLoading}
-                    >
-                      {newKeyLoading
-                        ? t("admin.creating")
-                        : t("admin.createKey")}
-                    </Button>
-                  </div>
-                </>
-              )}
-            </div>
-          )}
-          {apiKeys.map((key) => (
-            <div
-              key={key.id}
-              className="flex items-start justify-between py-2.5 border-b border-border last:border-0 gap-2"
-            >
-              <div className="flex flex-col gap-0.5 min-w-0 flex-1">
-                <div className="flex items-center gap-1.5 flex-wrap">
-                  <span className="text-xs font-semibold truncate">
-                    {key.name}
-                  </span>
-                  {!key.isActive && (
-                    <span className="text-[9px] font-semibold px-1 py-px border border-destructive/40 bg-destructive/10 text-destructive">
-                      {t("admin.revokedBadge")}
-                    </span>
-                  )}
-                </div>
-                <span className="text-[10px] text-muted-foreground">
-                  {t("admin.apiKeyUser")}: {key.username}
-                </span>
-                <span className="text-[10px] font-mono text-muted-foreground truncate">
-                  {key.tokenPrefix}…
-                </span>
-                <span className="text-[10px] text-muted-foreground">
-                  {key.createdAt.split("T")[0]} ·{" "}
-                  {key.expiresAt
-                    ? key.expiresAt.split("T")[0]
-                    : t("admin.apiKeyNoExpiry")}
-                </span>
-              </div>
-              <Button
-                variant="ghost"
-                size="icon"
-                className="size-6 text-muted-foreground hover:text-destructive shrink-0"
-                onClick={async () => {
-                  try {
-                    await deleteApiKey(key.id);
-                    setApiKeys((prev) => prev.filter((k) => k.id !== key.id));
-                    toast.success(
-                      t("admin.revokeKeySuccess", { name: key.name }),
-                    );
-                  } catch {
-                    toast.error(t("admin.revokeKeyFailed"));
-                  }
-                }}
-              >
-                <Trash2 className="size-3" />
-              </Button>
-            </div>
-          ))}
-        </div>
-      </AccordionSection>
+        apiKeys={apiKeys}
+        setApiKeys={setApiKeys}
+        loadApiKeys={loadApiKeys}
+        showCreateKey={showCreateKey}
+        setShowCreateKey={setShowCreateKey}
+        createdKeyToken={createdKeyToken}
+        setCreatedKeyToken={setCreatedKeyToken}
+        newKeyName={newKeyName}
+        setNewKeyName={setNewKeyName}
+        newKeyUserId={newKeyUserId}
+        setNewKeyUserId={setNewKeyUserId}
+        newKeyExpiry={newKeyExpiry}
+        setNewKeyExpiry={setNewKeyExpiry}
+        users={users}
+        handleCreateApiKey={handleCreateApiKey}
+        newKeyLoading={newKeyLoading}
+      />
 
-      {/* Create User Dialog */}
-      <Dialog open={createUserOpen} onOpenChange={setCreateUserOpen}>
-        <DialogContent className="w-[calc(100vw-2rem)] sm:max-w-lg">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-bold">
-              {t("admin.createUserTitle")}
-            </DialogTitle>
-            <DialogDescription className="text-xs text-muted-foreground">
-              {t("admin.createUserDesc")}
-            </DialogDescription>
-          </DialogHeader>
-          <div className="flex flex-col gap-4 mt-1">
-            <div className="flex flex-col gap-1.5">
-              <label className="text-xs font-semibold uppercase tracking-widest text-muted-foreground">
-                {t("admin.createUserUsername")}{" "}
-                <span className="text-accent-brand">*</span>
-              </label>
-              <Input
-                placeholder={t("admin.createUserEnterUsername")}
-                value={newUsername}
-                onChange={(e) => setNewUsername(e.target.value)}
-                onKeyDown={(e) => e.key === "Enter" && handleCreateUser()}
-              />
-            </div>
-            <div className="flex flex-col gap-1.5">
-              <label className="text-xs font-semibold uppercase tracking-widest text-muted-foreground">
-                {t("admin.createUserPassword")}{" "}
-                <span className="text-accent-brand">*</span>
-              </label>
-              <div className="relative">
-                <Input
-                  type={showNewPassword ? "text" : "password"}
-                  placeholder="Enter password"
-                  value={newPassword}
-                  onChange={(e) => setNewPassword(e.target.value)}
-                  onKeyDown={(e) => e.key === "Enter" && handleCreateUser()}
-                  className="pr-9"
-                />
-                <button
-                  onClick={() => setShowNewPassword((o) => !o)}
-                  className="absolute right-2.5 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-                >
-                  {showNewPassword ? (
-                    <EyeOff className="size-4" />
-                  ) : (
-                    <Eye className="size-4" />
-                  )}
-                </button>
-              </div>
-              <span className="text-xs text-muted-foreground">
-                {t("admin.createUserPasswordHint")}
-              </span>
-            </div>
-          </div>
-          <div className="flex justify-end gap-2 mt-2">
-            <Button
-              variant="ghost"
-              onClick={() => {
-                setCreateUserOpen(false);
-                setNewUsername("");
-                setNewPassword("");
-              }}
-            >
-              {t("common.cancel")}
-            </Button>
-            <Button
-              variant="outline"
-              className="border-accent-brand/40 text-accent-brand hover:bg-accent-brand/10 hover:text-accent-brand"
-              onClick={handleCreateUser}
-              disabled={createUserLoading}
-            >
-              {createUserLoading
-                ? t("admin.creating")
-                : t("admin.createUserSubmit")}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <AdminCreateUserDialog
+        open={createUserOpen}
+        onOpenChange={setCreateUserOpen}
+        newUsername={newUsername}
+        setNewUsername={setNewUsername}
+        newPassword={newPassword}
+        setNewPassword={setNewPassword}
+        showNewPassword={showNewPassword}
+        setShowNewPassword={setShowNewPassword}
+        handleCreateUser={handleCreateUser}
+        createUserLoading={createUserLoading}
+      />
 
-      {/* Edit User Dialog */}
-      <Dialog open={editUserOpen} onOpenChange={setEditUserOpen}>
-        <DialogContent className="w-[calc(100vw-2rem)] sm:max-w-lg">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-bold">
-              {t("admin.editUserTitle", { username: editUserTarget?.username })}
-            </DialogTitle>
-            <DialogDescription className="text-xs text-muted-foreground">
-              {t("admin.editUserDesc")}
-            </DialogDescription>
-          </DialogHeader>
-          {editUserTarget && (
-            <div className="flex flex-col gap-0 mt-1 divide-y divide-border">
-              <div className="grid grid-cols-1 md:grid-cols-2 gap-x-6 gap-y-2 py-3">
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-xs text-muted-foreground uppercase tracking-widest font-semibold">
-                    {t("admin.editUserUsername")}
-                  </span>
-                  <span className="text-sm font-semibold">
-                    {editUserTarget.username}
-                  </span>
-                </div>
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-xs text-muted-foreground uppercase tracking-widest font-semibold">
-                    {t("admin.editUserAuthType")}
-                  </span>
-                  <span className="text-sm font-semibold">
-                    {editUserTarget.isOidc && editUserTarget.passwordHash
-                      ? t("admin.authTypeDual")
-                      : editUserTarget.isOidc
-                        ? t("admin.authTypeOidc")
-                        : t("admin.authTypeLocal")}
-                  </span>
-                </div>
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-xs text-muted-foreground uppercase tracking-widest font-semibold">
-                    {t("admin.editUserAdminStatus")}
-                  </span>
-                  <span className="text-sm font-semibold">
-                    {editUserTarget.isAdmin
-                      ? t("admin.adminStatusAdministrator")
-                      : t("admin.adminStatusRegularUser")}
-                  </span>
-                </div>
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-xs text-muted-foreground uppercase tracking-widest font-semibold">
-                    {t("admin.editUserUserId")}
-                  </span>
-                  <span className="text-xs font-mono text-muted-foreground truncate">
-                    {editUserTarget.id}
-                  </span>
-                </div>
-              </div>
-              <div className="flex items-center justify-between py-3">
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-sm font-medium">
-                    {t("admin.userAdminAccess")}
-                  </span>
-                  <span className="text-xs text-muted-foreground">
-                    {t("admin.userAdminAccessDesc")}
-                  </span>
-                </div>
-                <AdminToggle
-                  on={editUserTarget.isAdmin}
-                  onToggle={() => handleToggleAdmin(editUserTarget)}
-                />
-              </div>
-              <div className="flex flex-col gap-2 py-3">
-                <span className="text-sm font-medium">
-                  {t("admin.userRoles")}
-                </span>
-                {editUserRolesLoading ? (
-                  <span className="text-xs text-muted-foreground">
-                    {t("newUi.sidebar.snippets.loading")}
-                  </span>
-                ) : (
-                  <>
-                    {editUserRoles.length > 0 && (
-                      <div className="flex flex-wrap gap-1.5">
-                        {editUserRoles.map((ur) => {
-                          const roleInfo = roles.find(
-                            (r) => r.id === ur.roleId,
-                          );
-                          const isSystem = roleInfo?.isSystem ?? false;
-                          return (
-                            <span
-                              key={ur.roleId}
-                              className="inline-flex items-center gap-1 text-[10px] font-semibold px-1.5 py-0.5 border border-accent-brand/40 bg-accent-brand/10 text-accent-brand"
-                            >
-                              {ur.roleDisplayName}
-                              {!isSystem && (
-                                <button
-                                  onClick={async () => {
-                                    try {
-                                      await removeRoleFromUser(
-                                        editUserTarget.id,
-                                        ur.roleId,
-                                      );
-                                      setEditUserRoles((prev) =>
-                                        prev.filter(
-                                          (r) => r.roleId !== ur.roleId,
-                                        ),
-                                      );
-                                    } catch {
-                                      toast.error(t("admin.removeRoleFailed"));
-                                    }
-                                  }}
-                                  className="hover:text-destructive ml-0.5"
-                                >
-                                  ×
-                                </button>
-                              )}
-                            </span>
-                          );
-                        })}
-                      </div>
-                    )}
-                    {roles.filter(
-                      (r) =>
-                        !r.isSystem &&
-                        !editUserRoles.some((ur) => ur.roleId === r.id),
-                    ).length > 0 && (
-                      <div className="flex flex-col gap-1">
-                        <span className="text-[10px] text-muted-foreground uppercase tracking-widest font-semibold">
-                          {t("admin.addRole")}
-                        </span>
-                        <div className="flex flex-wrap gap-1.5">
-                          {roles
-                            .filter(
-                              (r) =>
-                                !r.isSystem &&
-                                !editUserRoles.some((ur) => ur.roleId === r.id),
-                            )
-                            .map((r) => (
-                              <button
-                                key={r.id}
-                                onClick={async () => {
-                                  try {
-                                    await assignRoleToUser(
-                                      editUserTarget.id,
-                                      r.id,
-                                    );
-                                    setEditUserRoles((prev) => [
-                                      ...prev,
-                                      {
-                                        userId: editUserTarget.id,
-                                        roleId: r.id,
-                                        roleName: r.name,
-                                        roleDisplayName: r.displayName,
-                                        grantedBy: "",
-                                        grantedByUsername: "",
-                                        grantedAt: new Date().toISOString(),
-                                      },
-                                    ]);
-                                  } catch {
-                                    toast.error(t("admin.assignRoleFailed"));
-                                  }
-                                }}
-                                className="inline-flex items-center gap-1 text-[10px] font-semibold px-1.5 py-0.5 border border-border text-muted-foreground hover:border-accent-brand/40 hover:text-accent-brand transition-colors"
-                              >
-                                + {r.displayName}
-                              </button>
-                            ))}
-                        </div>
-                      </div>
-                    )}
-                    {editUserRoles.length === 0 &&
-                      roles.filter((r) => !r.isSystem).length === 0 && (
-                        <span className="text-xs text-muted-foreground">
-                          {t("admin.noCustomRoles")}
-                        </span>
-                      )}
-                  </>
-                )}
-              </div>
-              <div className="flex items-center justify-between py-3">
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-sm font-medium">
-                    {t("admin.revokeAllUserSessions")}
-                  </span>
-                  <span className="text-xs text-muted-foreground">
-                    {t("admin.revokeAllUserSessionsDesc")}
-                  </span>
-                </div>
-                <Button
-                  variant="outline"
-                  size="sm"
-                  className="border-destructive/40 text-destructive hover:bg-destructive/10 hover:text-destructive shrink-0 ml-8"
-                  onClick={() => handleRevokeUserSessions(editUserTarget.id)}
-                  disabled={editUserLoading}
-                >
-                  {t("admin.revoke")}
-                </Button>
-              </div>
-              <div className="flex flex-col gap-2 py-3">
-                <div className="flex items-start gap-2.5 border border-destructive/30 bg-destructive/5 px-3 py-2.5">
-                  <AlertCircle className="size-4 text-destructive shrink-0 mt-0.5" />
-                  <span className="text-xs text-destructive">
-                    {t("admin.deleteUserWarning")}
-                  </span>
-                </div>
-                <Button
-                  variant="outline"
-                  className="w-full border-destructive/40 text-destructive hover:bg-destructive/10 hover:text-destructive"
-                  disabled={editUserTarget.isAdmin || editUserLoading}
-                  onClick={handleDeleteEditUser}
-                >
-                  <Trash2 className="size-3.5" />
-                  {editUserLoading
-                    ? t("admin.deleting")
-                    : t("admin.deleteUser", {
-                        username: editUserTarget.username,
-                      })}
-                </Button>
-              </div>
-            </div>
-          )}
-        </DialogContent>
-      </Dialog>
+      <AdminEditUserDialog
+        open={editUserOpen}
+        onOpenChange={setEditUserOpen}
+        editUserTarget={editUserTarget}
+        editUserLoading={editUserLoading}
+        editUserRoles={editUserRoles}
+        editUserRolesLoading={editUserRolesLoading}
+        roles={roles}
+        setEditUserRoles={setEditUserRoles}
+        handleToggleAdmin={handleToggleAdmin}
+        handleRevokeUserSessions={handleRevokeUserSessions}
+        handleDeleteEditUser={handleDeleteEditUser}
+      />
 
-      {/* Link Account Dialog */}
-      <Dialog open={linkAccountOpen} onOpenChange={setLinkAccountOpen}>
-        <DialogContent className="w-[calc(100vw-2rem)] sm:max-w-lg">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-bold">
-              {t("admin.linkAccountTitle")}
-            </DialogTitle>
-            <DialogDescription className="text-xs text-muted-foreground">
-              {t("admin.linkAccountDesc", {
-                username: linkAccountTarget?.username,
-              })}
-            </DialogDescription>
-          </DialogHeader>
-          <div className="flex flex-col gap-4 mt-1">
-            <div className="flex items-start gap-2.5 border border-destructive/30 bg-destructive/5 px-3 py-2.5">
-              <AlertCircle className="size-4 text-destructive shrink-0 mt-0.5" />
-              <div className="flex flex-col gap-1 text-xs text-destructive">
-                <span>{t("admin.linkAccountWarningTitle")}</span>
-                <ul className="list-disc list-inside space-y-0.5 ml-1">
-                  <li>{t("admin.linkAccountEffect1")}</li>
-                  <li>{t("admin.linkAccountEffect2")}</li>
-                  <li>{t("admin.linkAccountEffect3")}</li>
-                </ul>
-              </div>
-            </div>
-            <div className="flex flex-col gap-1.5">
-              <label className="text-xs font-semibold uppercase tracking-widest text-muted-foreground">
-                {t("admin.linkAccountTargetUsername")}{" "}
-                <span className="text-accent-brand">*</span>
-              </label>
-              <Input placeholder={t("admin.linkAccountTargetPlaceholder")} />
-            </div>
-          </div>
-          <div className="flex justify-end gap-2 mt-2">
-            <Button variant="ghost" onClick={() => setLinkAccountOpen(false)}>
-              {t("common.cancel")}
-            </Button>
-            <Button
-              variant="outline"
-              className="border-destructive/40 text-destructive hover:bg-destructive/10 hover:text-destructive"
-            >
-              {t("admin.linkAccounts")}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <AdminLinkAccountDialog
+        open={linkAccountOpen}
+        onOpenChange={setLinkAccountOpen}
+        linkAccountTarget={linkAccountTarget}
+      />
     </div>
   );
 }
